+++ projects/design-angular-kit/src/lib/design-angular-kit.module.ts
@@ -12,6 +12,10 @@ import { TooltipDirective } from './tooltip/tooltip.directive';
 import { TooltipComponent } from './tooltip/tooltip.component';
 import { ButtonComponent } from './button/button.component';
 import { BadgeDirective } from './badge/badge.directive';
+import { CollapseDirective } from './collapse/collapse.directive';
+import { CollapseItemComponent } from './collapse/collapse-item.component';
+import { CollapseGroupComponent } from './collapse/collapse-group.component';
+import { FormInputComponent } from './form-input/form-input.component';
 import { DropdownComponent } from './dropdown/dropdown.component';
 import { DropdownItemComponent } from './dropdown/dropdown-item.component';
 import { DropdownHeaderComponent } from './dropdown/dropdown-header.component';
@@ -38,6 +42,10 @@ import { TabComponent } from './tabs/tab.component';
     ButtonComponent,
     TooltipDirective,
     TooltipComponent,
+    CollapseDirective,
+    CollapseItemComponent,
+    CollapseGroupComponent,
+    FormInputComponent,
     PopoverDirective,
     PopoverComponent,
     DropdownComponent,
@@ -58,6 +66,10 @@ import { TabComponent } from './tabs/tab.component';
     ProgressBarComponent,
     ButtonComponent,
     TooltipDirective,
+    CollapseDirective,
+    CollapseItemComponent,
+    CollapseGroupComponent,
+    FormInputComponent,
     PopoverDirective,
     DropdownComponent,
     DropdownItemComponent,