+++ src/app/collapse/collapse-example/collapse-example.component.ts
@@ -1,0 +1,32 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'it-collapse-example',
+  templateUrl: './collapse-example.component.html',
+  styleUrls: ['./collapse-example.component.scss']
+})
+export class CollapseExampleComponent {
+
+  isShown = false;
+
+  showTime = '';
+  shownTime = '';
+  hideTime = '';
+  hiddenTime = '';
+
+  logShow() {
+    this.showTime = new Date().toISOString();
+  }
+
+  logShown() {
+    this.shownTime = new Date().toISOString();
+  }
+
+  logHide() {
+    this.hideTime = new Date().toISOString();
+  }
+
+  logHidden() {
+    this.hiddenTime = new Date().toISOString();
+  }
+}