+++ src/app/collapse/collapse-examples/collapse-examples.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'it-collapse-examples',
+  templateUrl: './collapse-examples.component.html',
+  styleUrls: ['./collapse-examples.component.scss']
+})
+export class CollapseExamplesComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}