+++ src/app/collapse/collapse-group-example/collapse-group-example.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'it-collapse-group-example',
+  templateUrl: './collapse-group-example.component.html',
+  styleUrls: ['./collapse-group-example.component.scss']
+})
+export class CollapseGroupExampleComponent {
+  accordion = false;
+}